refactor(tasks): move things around
