--- packages/sanity/src/core/studio/tasks/components/list/styles.ts
@@ -1,11 +1,0 @@
-import {Card, CardProps} from '@sanity/ui'
-import styled from 'styled-components'
-
-export const ThreadCard = styled(Card).attrs<CardProps>(({tone}) => ({
-  padding: 3,
-  radius: 3,
-  sizing: 'border',
-  tone: tone || 'transparent',
-}))<CardProps>`
-  // ...
-`
