+++ packages/sanity/src/core/studio/tasks/components/sidebar/index.ts
